Using IModularTenantEvents
