--- src/OrchardCore.Modules/OrchardCore.Workflows/Http/Services/IWorkflowRoutesProvider.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OrchardCore.Workflows.Http.Services
-{
-    public interface IWorkflowRoutesProvider
-    {
-        Task RegisterRoutesAsync();
-    }
-}
