+++ src/OrchardCore.Modules/OrchardCore.Workflows/Http/Startup.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
@@ -10,7 +8,6 @@ using OrchardCore.Modules;
 using OrchardCore.Scripting;
 using OrchardCore.Workflows.Helpers;
 using OrchardCore.Workflows.Http.Activities;
-using OrchardCore.Workflows.Http.Controllers;
 using OrchardCore.Workflows.Http.Drivers;
 using OrchardCore.Workflows.Http.Filters;
 using OrchardCore.Workflows.Http.Handlers;
@@ -18,7 +15,6 @@ using OrchardCore.Workflows.Http.Liquid;
 using OrchardCore.Workflows.Http.Scripting;
 using OrchardCore.Workflows.Http.Services;
 using OrchardCore.Workflows.Http.WorkflowContextProviders;
-using OrchardCore.Workflows.Models;
 using OrchardCore.Workflows.Services;
 
 namespace OrchardCore.Workflows.Http
@@ -54,8 +50,7 @@ namespace OrchardCore.Workflows.Http
             services.AddScoped<ILiquidTemplateEventHandler, SignalLiquidTemplateHandler>();
             services.AddLiquidFilter<SignalUrlFilter>("signal_url");
 
-            services.AddScoped<IWorkflowRoutesProvider, WorkflowTypeRoutesProvider>();
-            services.AddScoped<IWorkflowRoutesProvider, HttpRequestFilterRouteProvider>();
+            services.AddTransient<IModularTenantEvents, HttpRequestRouteActivator>();
         }
 
         public override void Configure(IApplicationBuilder app, IRouteBuilder routes, IServiceProvider serviceProvider)
@@ -73,12 +68,6 @@ namespace OrchardCore.Workflows.Http
                 template: "workflows/invoke/{token}",
                 defaults: new { controller = "HttpWorkflow", action = "Invoke" }
             );
-
-            var workflowRoutesProviders = serviceProvider.GetServices<IWorkflowRoutesProvider>();
-            foreach (var workflowRoutesProvider in workflowRoutesProviders)
-            {
-                workflowRoutesProvider.RegisterRoutesAsync().GetAwaiter().GetResult();
-            }
         }
     }
 }
