+++ src/OrchardCore/OrchardCore.Abstractions/Modules/IModularTenantEvents.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading.Tasks;
+using System.Threading.Tasks;
 
 namespace OrchardCore.Modules
 {
@@ -8,5 +8,28 @@ namespace OrchardCore.Modules
         Task ActivatedAsync();
         Task TerminatingAsync();
         Task TerminatedAsync();
+    }
+
+    public class ModularTenantEvents : IModularTenantEvents
+    {
+        public virtual Task ActivatedAsync()
+        {
+            return Task.CompletedTask;
+        }
+
+        public virtual Task ActivatingAsync()
+        {
+            return Task.CompletedTask;
+        }
+
+        public virtual Task TerminatedAsync()
+        {
+            return Task.CompletedTask;
+        }
+
+        public virtual Task TerminatingAsync()
+        {
+            return Task.CompletedTask;
+        }
     }
 }
